+++ includes/Activator.php
@@ -15,13 +15,6 @@ defined( 'WPINC' ) || die;
 class Activator {
 
 	public static function activate() {
-
-        // Test is it working
-        update_option( '__activate', time() );
-
-        error_log( 'Activator' );
-        error_log( function_exists('update_option') );
-
-        return 'activate';
+        // Put code that you want to run on activation in here
 	}
 }
