+++ includes/I18n.php
@@ -17,8 +17,11 @@ class I18n {
 
 	/**
 	 * Load the plugin text domain for translation.
+	 * Text domain always needs to be passed as a string
+	 * More info on link below
 	 *
 	 * @since    1.0.0
+	 * @link ( https://developer.wordpress.org/themes/functionality/internationalization/#add-text-domain-to-strings, Add text domain to strings )
 	 */
 	public function load_plugin_textdomain() {
