+++ includes/admin/Admin.php
@@ -2,7 +2,7 @@
 
 namespace wppb\admin;
 
-use wppb\Assets;
+use wppb\Utils;
 use wppb\status\System_Report;
 use wppb\status\System_Options;
 
@@ -18,26 +18,35 @@ defined( 'WPINC' ) || die;
  */
 class Admin {
 
+	use Utils;
+
 	/**
 	 * Initialize the class and set its properties.
 	 *
 	 * @since    1.0.0
 	 */
 	public function __construct() {
-		// Load assets from manifest.json
-		$this->assets = new Assets();
-		add_action( 'admin_menu', array( __CLASS__, 'register_menu_page' ) );
-		add_action( 'admin_menu', array( __CLASS__, 'register_submenu_page' ) );
+
+		add_action( 'admin_menu', array( $this, 'register_menu_page' ) );
 	}
 
 	/**
-	 * Register the stylesheets for the admin area.
+	 * Enqueue the stylesheets for wp-admin
 	 *
 	 * @since    1.0.0
 	 */
 	public function enqueue_styles() {
 
-		wp_enqueue_style( CS_WPPB_PLUGIN_NAME . '/wp/css', $this->assets->get( 'styles/admin.css' ), array(), CS_WPPB_PLUGIN_VERSION, 'all' );
+		// Generate wp-admin CSS handle
+		$handle = $this->get_plugin_id() . '/wp/css';
+
+		wp_enqueue_style(
+			$handle,
+			$this->get_asset( 'styles/admin.css' ),
+			array(),
+			$this->get_plugin_version(),
+			'all'
+		);
 	}
 
 	/**
@@ -47,54 +56,53 @@ class Admin {
 	 */
 	public function enqueue_scripts() {
 
+		// Generate wp-admin Js handle
+		$handle = $this->get_plugin_id() . '/wp/js';
+
 		wp_enqueue_script(
-			CS_WPPB_PLUGIN_NAME . '/wp/js', $this->assets->get( 'scripts/admin.js' ), array(), CS_WPPB_PLUGIN_VERSION, false );
-		
-		wp_enqueue_script( CS_WPPB_PLUGIN_NAME . '/wp/js-clipboard', $this->assets->get( 'scripts/clipboard.js' ), array(), '2.0', true );
+			$handle,
+			$this->get_asset( 'scripts/admin.js' ),
+			array(),
+			$this->get_plugin_version(),
+			false
+		);
 	}
 
 	/**
-	 * Register menu page.
-	 * Only viewable by Administrators
+	 * Register menu and submenu page
+	 * Only accessible for user with manage_options capability
 	 *
 	 * @since  1.0.1
 	 * @action admin_menu
 	 *
 	 * @return void
 	 */
-	public static function register_menu_page() {
+	public function register_menu_page() {
 
+		// Register menu page
+		$page_id = $this->get_plugin_id();
+
+		// Menu item
 		add_menu_page(
 			__( 'Plugin Settings', 'cs-wppb' ),
 			__( 'Plugin Settings', 'cs-wppb' ),
 			'manage_options',
-			'cs-wppb',
-			array( __CLASS__, 'render_options' ),
+			$page_id,
+			array( $this, 'render_options' ),
 			'dashicons-admin-generic'
 		);
 
-	}
-
-	/**
-	 * Register submenu page and enqueue styles and scripts.
-	 * Only viewable by Administrators
-	 *
-	 * @since  1.0.1
-	 * @action admin_menu
-	 *
-	 * @return void
-	 */
-	public static function register_submenu_page() {
-
+		// Submenu item
 		add_submenu_page(
-			'cs-wppb',
+			$page_id,
 			__( 'System Status', 'cs-wppb' ),
 			__( 'System Status', 'cs-wppb' ),
 			'manage_options',
-			'cs-wppb-system-status',
-			array( __CLASS__, 'render_info' ),
+			$page_id . '-system-status',
+			array( $this, 'render_info' ),
 		);
 	}
+
 
 	/**
 	 * Render plugin system status
@@ -103,8 +111,8 @@ class Admin {
 	 *
 	 * @return void
 	 */
-	public static function render_info() {
-		System_Report::system_report();
+	public function render_info() {
+		// System_Report::system_report();
 	}
 
 	/**
@@ -114,8 +122,8 @@ class Admin {
 	 *
 	 * @return void
 	 */
-	public static function render_options() {
-		System_Options::system_options();
+	public function render_options() {
+		// System_Options::system_options();
 	}
 
 }
