+++ src/config.js
@@ -1,15 +1,29 @@
 module.exports = {
     entry: {
-        main: ["scripts/main.js", "styles/main.scss"],		// root javascript and style files - frontend
-        admin: ["scripts/admin.js", "styles/admin.scss"],	// root javascript and style files - backend
-        clipboard: ["scripts/_npm-modules/clipboard.js"],	// clipboard for c/p system status report
+        main: [
+            "scripts/main.js",
+            "styles/main.scss"
+        ],
+        admin: [
+            "scripts/admin.js",
+            "styles/admin.scss",
+        ],
     },
-    openBrowserOnWatch: false,				// open browser with devUrl url on watch
-    showErrorsInBrowser: true,				// show webpack errors on frontent
-    useSSLinDev: false,						// use https: in devUrl
-    publicPath: "/wp-content/plugins/wppb",	// path to plugin
-    devUrl: "http://mylocal.domain",		// localhost or custom domain name
-    proxyUrl: "http://localhost",			// proxyUrl
-    proxyPort: 3000,						// port to use on "watch"
-    watch: ["includes/**/*.php"],			// folders / files to include in live reload
+    // open browser with devUrl url when watch mode starts
+    openBrowserOnWatch: false,
+    // show webpack compailing errors in browser while in watch mode
+    showErrorsInBrowser: true,
+    // use https: in devUrl
+    useSSLinDev: false,
+    // path to plugin folder
+    publicPath: "/wp-content/plugins/wppb",
+    // localhost or custom host name
+    devUrl: "http://mylocal.domain",
+    // proxyUrl, when watch mode is enabled you will access it on this url
+    proxyUrl: "http://localhost",
+    // proxy port used in watch mode
+    proxyPort: 3000,
+    // folders / files to include in while in watch mode to monitor for changes
+    // change in css/js is automatically included
+    watch: ["includes/**/*.php"],
 };
