+++ src/styles/_utils/_variables.scss
@@ -13,7 +13,7 @@ $grid: (
 $fonts: (
     "base-size": 16,
     "helvetica": "Helvetica, sans-serif",
-    "graphik": "Graphik, Arial, sans-serif",
+    "arial": "Arial, sans-serif",
 );
 
 /**
